refactored recent projects actions
