+++ src/js/components/home/projectsManagement/MyProjects.js
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import ProjectCard from './ProjectCard';
 
 let MyProjects = (props) => {
-  const {myProjects} = props;
+  const {myProjects, user} = props;
   const projects = myProjects.map( (projectDetails, index) =>
-    <ProjectCard key={index} projectDetails={projectDetails} actions={props.actions} />
+    <ProjectCard user={user} key={index} projectDetails={projectDetails} actions={props.actions} />
   );
 
   return (
